feat(TDP-3845): create interfaces for export strategies
- make TcompExportStrategy available for sampling
- make Streams export strategy available for S3 if streams available and compatible
