+++ dataprep-transformation/src/main/java/org/talend/dataprep/transformation/service/export/SampleExportStrategy.java
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2006-2017 Talend Inc. - www.talend.com
+ *
+ * This source code is available under agreement available at
+ * https://github.com/Talend/data-prep/blob/master/LICENSE
+ *
+ * You should have received a copy of the agreement
+ * along with this program; if not, write to Talend SA
+ * 9 rue Pages 92150 Suresnes, France
+ */
+
+package org.talend.dataprep.transformation.service.export;
+
+import org.talend.dataprep.transformation.service.ExportStrategy;
+
+/**
+ * Tagging interface for export strategies that can be used for sampling.
+ */
+public interface SampleExportStrategy extends ExportStrategy {}
